Format code

+++ Assets/Scripts/Modules/SpriteSheetAnimations/SpriteSheetAnimation.cs
@@ -1,25 +1,21 @@
 using UnityEngine;
 
-namespace Metroidvania.Animations
-{
+namespace Metroidvania.Animations {
     [System.Serializable]
-    public class SpriteSheetAnimation : ISerializationCallbackReceiver
-    {
+    public class SpriteSheetAnimation : ISerializationCallbackReceiver {
         public string name;
         [HideInInspector] public int hash;
         public int frameRate = 12;
         public bool loop;
         public Sprite[] sheet;
 
-        public float duration => (float)sheet.Length / (float)frameRate;
+        public float duration => sheet.Length / (float)frameRate;
 
-        public void OnAfterDeserialize()
-        {
+        public void OnAfterDeserialize() {
             hash = Animator.StringToHash(name);
         }
 
-        public void OnBeforeSerialize()
-        {
+        public void OnBeforeSerialize() {
         }
     }
 }
